Windows CE performance patches ported over from zeromq4-x.
The patches allow using ZeroMQ on Windows CE 4.2, in which several
socket functions are either buggy or inexistent. Additionally,
the signaler is modified to use another mechanism than select(),
which is extremely slow under Windows CE, probably due to poor driver
design and the lack of a loopback IP adapter. On systems with slow
network interfaces, the modified signaler improves inproc latency
by over an order of magnitude.